add integrations tests and refactor code to make it testable

+++ api/models/weather.go
@@ -5,8 +5,8 @@ import "gorm.io/gorm"
 type Weather struct {
 	gorm.Model
 	RecordedAt  string
-	Humidity    float32
-	Temperature float32
+	Humidity    float64
+	Temperature float64
 }
 
 func (w Weather) TableName() string {
